Initial commit
Adding initial commit with ConfigParser from Storm and Paramiko. Next
commit will be for removing unnecessary parts from both and try and
make it as lean as possible, as well as add some tests.

Also needs to get cleaned up as this is just bare minimum to get
it all working.